Primitive class implementation
